ROOT macros to read the FoCalH data with minor changes to the code to avoid warnings

+++ reader/ProtoFocalEvent.hh
@@ -31,7 +31,7 @@ struct ProtoFoCalHEvent{
   float energy;
   float time;
   float charge;
-  float evn;
+  int evn;
   int t0event;
   int t1event;
   int bOK[2];
